OMP tasks 1-9

+++ parallel_computing/omp/src/sections.cpp
@@ -6,14 +6,11 @@
 #include <omp.h>
 #include <ctime>
 
-int main()
-{
+int main() {
     int d[6][8];
     std::srand(time(nullptr));
-    for (int i = 0; i < 6; i++)
-    {
-        for (int j = 0; j < 8; j++)
-        {
+    for (int i = 0; i < 6; i++) {
+        for (int j = 0; j < 8; j++) {
             d[i][j] = std::rand() % 10;
         }
     }
@@ -22,26 +19,22 @@ int main()
     {
 #pragma omp section
         {
-            int avg = 0;
-            for (int i = 0; i < 6; i++)
-            {
-                for (int j = 0; j < 8; j++)
-                {
+            double avg = 0;
+            for (int i = 0; i < 6; i++) {
+                for (int j = 0; j < 8; j++) {
                     avg += d[i][j];
                 }
             }
             avg /= 48;
-            std::printf("Average = %d\t(c) Thread %d\n", avg, omp_get_thread_num());
+            std::printf("Average = %lf\t(c) Thread %d\n", avg, omp_get_thread_num());
 
         }
 #pragma omp section
         {
             int min = d[0][0];
             int max = d[0][0];
-            for (int i = 0; i < 6; i++)
-            {
-                for (int j = 0; j < 8; j++)
-                {
+            for (int i = 0; i < 6; i++) {
+                for (int j = 0; j < 8; j++) {
                     if (max < d[i][j])
                         max = d[i][j];
                     else if (min > d[i][j])
@@ -53,10 +46,8 @@ int main()
 #pragma omp section
         {
             int count3 = 0;
-            for (int i = 0; i < 6; i++)
-            {
-                for (int j = 0; j < 8; j++)
-                {
+            for (int i = 0; i < 6; i++) {
+                for (int j = 0; j < 8; j++) {
                     if (d[i][j] % 3 == 0)
                         count3++;
                 }
